requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
supabase==2.4.0
httpx==0.25.2
realtime==1.0.2
gotrue==2.4.2
bcrypt>=4.0.0
